feat(rancher/aws): Use AWS env vars, use null defaults
By using null defaults for AWS credentials, the Provider will use Env vars that are set and automatically leverage temporary credentials as well. A user can always decide to set the values like before. But by default it will 'just work' if there are already environment variables set.

The availability zone now also defaults to null, which leaves the decision which one to deploy the resources into up to the Provider. If the variable is set, the behavior is like before